Add a rescue configuration for break-fix apply
This commit provides a database rescue configuration that can be quickly
applied to restore broken pds-server access to PostgreSQL, in the event
that it is removed.

This is necessary because if the pds trusted-external-command is
configured but pds-server is non-functional, then a normal Puppet agent
run cannot complete.

The rescue class should restore enough access that a `puppet agent -t`
run can proceed.